Moved wine queries to their own file and updated resolvers
