+++ server/schemas/queries/index.js
@@ -2,3 +2,4 @@ export { getAllUsers, getSpecificUser, getMe } from "./user.js";
 export { getVarieties } from "./variety.js";
 export { getRegions } from "./region.js";
 export { getProducers } from "./producer.js";
+export { searchWine, getSpecificWine } from "./wine.js";
